refactor container registry properties editor
