+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-containerregistry/src/main/java/com/microsoft/azure/toolkit/intellij/containerregistry/properties/ContainerRegistryPropertiesEditorProvider.java
@@ -3,7 +3,7 @@
  * Licensed under the MIT License. See License.txt in the project root for license information.
  */
 
-package com.microsoft.azure.toolkit.intellij.legacy.docker;
+package com.microsoft.azure.toolkit.intellij.containerregistry.properties;
 
 import com.intellij.openapi.fileEditor.FileEditor;
 import com.intellij.openapi.fileEditor.FileEditorPolicy;
@@ -11,35 +11,35 @@ import com.intellij.openapi.fileEditor.FileEditorProvider;
 import com.intellij.openapi.project.DumbAware;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.vfs.VirtualFile;
-
 import com.microsoft.azure.toolkit.lib.common.messager.ExceptionNotification;
 import com.microsoft.azure.toolkit.lib.common.operation.AzureOperation;
-import org.jetbrains.annotations.NotNull;
 
-public class ContainerRegistryPropertyViewProvider implements FileEditorProvider, DumbAware {
+import javax.annotation.Nonnull;
 
-    public static final String TYPE = "CONTAINER_REGISTRY_PROPERTY_VIEW";
+public class ContainerRegistryPropertiesEditorProvider implements FileEditorProvider, DumbAware {
+
+    public static final String CONTAINER_REGISTRY_PROPERTY_TYPE = "Microsoft.ContainerRegistry.registries";
 
     @Override
-    public boolean accept(@NotNull Project project, @NotNull VirtualFile virtualFile) {
-        return virtualFile.getFileType().getName().equals(TYPE);
+    public boolean accept(@Nonnull Project project, @Nonnull VirtualFile virtualFile) {
+        return virtualFile.getFileType().getName().equals(getEditorTypeId());
     }
 
-    @NotNull
+    @Nonnull
     @Override
     @ExceptionNotification
     @AzureOperation(name = "platform/acr.create_properties_editor.registry", params = {"virtualFile.getName()"})
-    public FileEditor createEditor(@NotNull Project project, @NotNull VirtualFile virtualFile) {
-        return new ContainerRegistryPropertyView(project, virtualFile);
+    public FileEditor createEditor(@Nonnull Project project, @Nonnull VirtualFile virtualFile) {
+        return new ContainerRegistryPropertiesEditor(project, virtualFile);
     }
 
-    @NotNull
+    @Nonnull
     @Override
     public String getEditorTypeId() {
-        return TYPE;
+        return CONTAINER_REGISTRY_PROPERTY_TYPE;
     }
 
-    @NotNull
+    @Nonnull
     @Override
     public FileEditorPolicy getPolicy() {
         return FileEditorPolicy.HIDE_DEFAULT_EDITOR;
