+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-containerregistry/src/main/resources/META-INF/azure-intellij-plugin-containerregistry.xml
@@ -6,6 +6,6 @@
     </extensions>
 
     <extensions defaultExtensionNs="com.intellij">
-        <fileEditorProvider implementation="com.microsoft.azure.toolkit.intellij.legacy.docker.ContainerRegistryPropertyViewProvider"/>
+        <fileEditorProvider implementation="com.microsoft.azure.toolkit.intellij.containerregistry.properties.ContainerRegistryPropertiesEditorProvider"/>
     </extensions>
 </idea-plugin>
